# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: LFX V2 - Survey Service
    description: Proxy service for ITX survey system
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /surveys/{survey_id}:
        get:
            tags:
                - survey
            summary: get_survey survey
            description: |-
                Get survey details (proxies to ITX GET /v2/surveys/{survey_id})

                **Required security scopes for jwt**:
                  * `manage:projects`
                  * `manage:surveys`
            operationId: survey#get_survey
            parameters:
                - name: survey_id
                  in: path
                  description: Survey identifier
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: JWT token
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SurveyScheduleResult'
                        required:
                            - id
                            - survey_status
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/UnauthorizedError'
                        required:
                            - code
                            - message
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/ForbiddenError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        put:
            tags:
                - survey
            summary: update_survey survey
            description: |-
                Update survey (proxies to ITX PUT /v2/surveys/{survey_id}). Only allowed when status is 'disabled'

                **Required security scopes for jwt**:
                  * `manage:projects`
                  * `manage:surveys`
            operationId: survey#update_survey
            parameters:
                - name: survey_id
                  in: path
                  description: Survey identifier
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: JWT token
                  required: false
                  type: string
                - name: update_survey_request_body
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SurveyUpdateSurveyRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SurveyScheduleResult'
                        required:
                            - id
                            - survey_status
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/UnauthorizedError'
                        required:
                            - code
                            - message
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/ForbiddenError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        delete:
            tags:
                - survey
            summary: delete_survey survey
            description: |-
                Delete survey (proxies to ITX DELETE /v2/surveys/{survey_id}). Only allowed when status is 'disabled'

                **Required security scopes for jwt**:
                  * `manage:projects`
                  * `manage:surveys`
            operationId: survey#delete_survey
            parameters:
                - name: survey_id
                  in: path
                  description: Survey identifier
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: JWT token
                  required: false
                  type: string
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/UnauthorizedError'
                        required:
                            - code
                            - message
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/ForbiddenError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /surveys/schedule:
        post:
            tags:
                - survey
            summary: schedule_survey survey
            description: |-
                Create a scheduled survey for ITX project committees (proxies to ITX POST /surveys/schedule)

                **Required security scopes for jwt**:
                  * `manage:projects`
                  * `manage:surveys`
            operationId: survey#schedule_survey
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token
                  required: false
                  type: string
                - name: schedule_survey_request_body
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SurveyScheduleSurveyRequestBody'
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/SurveyScheduleResult'
                        required:
                            - id
                            - survey_status
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - code
                            - message
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/UnauthorizedError'
                        required:
                            - code
                            - message
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/ForbiddenError'
                        required:
                            - code
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - code
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - code
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - code
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - code
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
definitions:
    BadRequestError:
        title: BadRequestError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Eius consequatur sint nemo voluptatibus natus.
            message:
                type: string
                description: Error message
                example: Ipsa cum dolore aut aut earum.
        description: Bad request
        example:
            code: Nulla voluptates.
            message: Molestiae officia amet consequatur est cupiditate.
        required:
            - code
            - message
    ConflictError:
        title: ConflictError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Voluptatum sed temporibus mollitia eum omnis.
            message:
                type: string
                description: Error message
                example: Sit quaerat pariatur.
        description: Conflict
        example:
            code: Quibusdam minima ut quis dolore incidunt velit.
            message: Velit eum consectetur inventore.
        required:
            - code
            - message
    ForbiddenError:
        title: ForbiddenError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Ut provident qui voluptas explicabo.
            message:
                type: string
                description: Error message
                example: Eos eum exercitationem repellat est id rerum.
        description: Forbidden
        example:
            code: Enim veritatis quidem dolorum eos qui id.
            message: Id illo sunt error accusamus.
        required:
            - code
            - message
    InternalServerError:
        title: InternalServerError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Nihil corporis ad est.
            message:
                type: string
                description: Error message
                example: Harum molestiae ex explicabo.
        description: Internal server error
        example:
            code: Cum voluptatem eius.
            message: Occaecati modi iure.
        required:
            - code
            - message
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Nemo a natus eius fugiat.
            message:
                type: string
                description: Error message
                example: Sunt ipsa incidunt ullam ad nulla vitae.
        description: Not found
        example:
            code: Deleniti occaecati numquam.
            message: Temporibus voluptas quidem sint omnis et.
        required:
            - code
            - message
    ServiceUnavailableError:
        title: ServiceUnavailableError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Sit accusantium quasi illum quasi ratione minus.
            message:
                type: string
                description: Error message
                example: Libero sit rerum voluptatem dolorem est.
        description: Service unavailable
        example:
            code: Magnam reiciendis voluptas eaque culpa illo.
            message: Ad non.
        required:
            - code
            - message
...
    UnauthorizedError:
        title: UnauthorizedError
        type: object
        properties:
            code:
                type: string
                description: HTTP status code
                example: Culpa minus consequuntur vitae.
            message:
                type: string
                description: Error message
                example: Et in possimus nihil et aut explicabo.
        description: Unauthorized
        example:
            code: Iusto nemo impedit doloremque deleniti.
            message: Sint at quaerat.
        required:
            - code
            - message
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: |-
            Heimdall JWT authorization

            **Security Scopes**:
              * `read:projects`: Read project data
              * `manage:projects`: Manage projects
              * `manage:surveys`: Manage surveys
        name: Authorization
        in: header